Move methods in `BoxLayoutWidget` to `RenderWidget`, Add SizedBox
