+++ examples/boxes.rs
@@ -10,10 +10,14 @@ struct App;
 impl ViewWidget for App {
     fn build<'w>(&'w self, _: BuildContext<'w, Self>) -> Self::Widget<'w> {
         UnconstrainedBox {
-            child: ColoredBox {
-                child: Text::new("Hello world!"),
-                color: Color::RED,
-            },
+            child: SizedBox::new(
+                ColoredBox {
+                    child: Text::new("Hello world!"),
+                    color: Color::RED,
+                },
+                Some(100.0),
+                Some(100.0)
+            )
         }
     }
 }
